Id incremental changes

+++ Account/src/main/java/com/sept/challenge2/account/service/AccountService.java
@@ -16,9 +16,14 @@ public class AccountService {
     AccountDao accountDao;
 
     public Account createAccount(Account account) {
-        int numOfAccounts = accountDao.findAll().size();
-        account.setAccountNumber(numOfAccounts+1);
-        return accountDao.save(account);
+        List<Account> numOfAccounts = accountDao.findAll();
+        for (int i=0; i<numOfAccounts.size();i++){
+            if (i==numOfAccounts.size()-1){
+                account.setAccountNumber(numOfAccounts.get(i).getAccountNumber()+1);
+                return accountDao.save(account);
+            }
+        }
+        return null;
     }
 
     public Optional getAccountById(Integer id) throws NoSuchElementException{
@@ -52,9 +57,14 @@ public class AccountService {
     }
 
     public Account deleteAccount(Account account){
+        System.out.println(account.getAccountNumber());
+        System.out.println(account.getId());
         List<Account> allAccounts = accountDao.findAll();
         for (int i=0; i<allAccounts.size();i++){
-            if (allAccounts.get(i).getId()==account.getId()) {
+            System.out.println(allAccounts.get(i).getAccountNumber());
+            System.out.println(allAccounts.get(i).getId());
+            if (allAccounts.get(i).getAccountNumber()==account.getAccountNumber() &&
+                    allAccounts.get(i).getId()==account.getId()) {
                 accountDao.delete(allAccounts.get(i));
                 return allAccounts.get(i);
             }
